CRUD Pages Done

+++ Backend/src/main/java/com/example/demo/repository/AuthRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.Auth;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthRepository extends JpaRepository<Auth, Long> {
+public interface AuthRepository extends JpaRepository<Auth, Integer> {
 }
